fix: add logging to empty catch blocks and suppress deprecation warnings
Address code quality issues identified during
project audit to improve reliability and
reduce compiler warnings.

- Add SLF4J logging to 19 empty catch blocks
  across 10 files to prevent silent exception
  swallowing
- Replace deprecated NetcdfFile.getDimensions()
  with getRootGroup().getDimensions()
- Replace deprecated Attribute(String, Array)
  constructor with Attribute.builder() pattern
- Suppress deprecation warnings for
  NetcdfFileWriter usage pending full migration
- Remove 2 redundant casts in HDF5 adapters
- Add missing serialVersionUID to HDF5Exception

+++ src/main/java/com/github/neuroinformatics/samurai_graph/lib/hdf5/Hdf5ReaderAdapter.java
@@ -413,7 +413,7 @@ class Hdf5ReaderAdapter implements IHDF5Reader {
         Attribute attr = node.getAttribute(attributeName);
         int numElements = (int) attr.getSize();
         return new HDF5DataTypeInformation(
-            mapDataClass(attr.getDataType().getDataClass()), (int) numElements);
+            mapDataClass(attr.getDataType().getDataClass()), numElements);
       } catch (Exception e) {
         logger.error("Failed to get attribute info: {}@{}", attributeName, path, e);
         throw new RuntimeException(e);
